update: update pipeline

+++ dataRawProcess/audio_processing/rating_audio.py
@@ -5,22 +5,37 @@ from concurrent.futures import ThreadPoolExecutor
 from wvmos import get_wvmos
 from utils.logger import Logger
 import pandas as pd
+from utils.tool import check_exists
 
-logger = Logger()
+logger = Logger.get_logger()
 model = get_wvmos()
 
 def run_predict(audio_folder_path, playlist_name, episode):
+
     os.makedirs(f'/home4/tuanlha/EXpressiveTTS/dataRawProcess/05_data_extract/{playlist_name}', exist_ok=True)
+    if os.path.exists(f'/home4/tuanlha/EXpressiveTTS/dataRawProcess/05_data_extract/{playlist_name}/{episode}_NISQA_results.csv'):
+        logger.info(f"File {episode}_wvmos.csv already exists, skipping...")
+        return
     # os.makedirs(f'/home4/tuanlha/EXpressiveTTS/dataRawProcess/05_data_extract/{playlist_name}/{episode}', exist_ok=True)
-    command = f"python run_predict.py --mode predict_dir --pretrained_model weights/nisqa.tar --data_dir '{audio_folder_path}' --num_workers 8 --bs 30 --output_dir /home4/tuanlha/EXpressiveTTS/dataRawProcess/05_data_extract/{playlist_name}"
-    print(command)
-    subprocess.run(command, shell=True, cwd="/home4/tuanlha/NISQA")
+    # command = f"python run_predict.py --mode predict_dir --pretrained_model weights/nisqa.tar --data_dir '{audio_folder_path}' --num_workers 8 --bs 30 --output_dir /home4/tuanlha/EXpressiveTTS/dataRawProcess/05_data_extract/{playlist_name}"
+    command = [
+        "python", "run_predict.py",
+        "--mode", "predict_dir",
+        "--pretrained_model", "weights/nisqa.tar",
+        "--data_dir", audio_folder_path,
+        "--num_workers", "8",
+        "--bs", "30",
+        "--output_dir", f"/home4/tuanlha/EXpressiveTTS/dataRawProcess/05_data_extract/{playlist_name}"
+    ]
+
+    print(" ".join(command))
+    subprocess.run(command, cwd="/home4/tuanlha/NISQA")
 def limit_cpu_for_diarization():
     # return
     """Giới hạn process và các subprocess chỉ chạy trên core 0 và 1"""
     try:
         p = psutil.Process(os.getpid())
-        p.cpu_affinity([i for i in range(36,40)])  # Chỉ dùng core 0 và 1
+        p.cpu_affinity([i for i in range(34,40)])  # Chỉ dùng core 0 và 1
         print(f"Process {os.getpid()} bị giới hạn trên core: {p.cpu_affinity()}")
     except Exception as e:
         print(f"Không thể thiết lập cpu_affinity: {e}")
@@ -31,8 +46,11 @@ def rating_audio(args, cfg):
     episode_list = sorted(os.listdir(os.path.join(args.data_path, args.playlist_name)))
     episode_name = [os.path.basename(ep).rsplit('.', 1)[0] for ep in episode_list]
 
-    audio_folder_path = f'/home4/tuanlha/EXpressiveTTS/dataRawProcess/04_vad/{playlist_name}'
+    audio_folder_path = f'/home4/tuanlha/EXpressiveTTS/dataRawProcess/04_denoise/{playlist_name}'
     os.makedirs(audio_folder_path, exist_ok=True)
+
+
+
     with ThreadPoolExecutor(max_workers=2, initializer=limit_cpu_for_diarization) as executor:
         limit_cpu_for_diarization()
         for cnt, episode in enumerate(episode_name):
@@ -40,17 +58,21 @@ def rating_audio(args, cfg):
     # logger.info("Done rating with NISQA")
     # dct = {}
     for episode in episode_name:
-
-        mos = model.calculate_dir(os.path.join(audio_folder_path, episode), mean=False)
-        file_list = sorted(os.listdir(os.path.join(audio_folder_path, episode)))
-        # file_list = [os.path.join(audio_folder_path, episode, file) for file in file_list]
-        # print(file_list)
-        df = pd.DataFrame({
-            'audio_path': file_list,  # Cột 1: đường dẫn file
-            'score': mos      # Cột 2: điểm số tương ứng
-        })
-        df.to_csv(f'/home4/tuanlha/EXpressiveTTS/dataRawProcess/05_data_extract/{playlist_name}/{episode}_wvmos.csv', index=False)
-
+        if os.path.exists(f'/home4/tuanlha/EXpressiveTTS/dataRawProcess/05_data_extract/{playlist_name}/{episode}_wvmos.csv'):
+            continue
+        try:
+            mos = model.calculate_dir(os.path.join(audio_folder_path, episode), mean=False)
+            file_list = sorted(os.listdir(os.path.join(audio_folder_path, episode)))
+            # file_list = [os.path.join(audio_folder_path, episode, file) for file in file_list]
+            # print(file_list)
+            df = pd.DataFrame({
+                'audio_path': file_list,  # Cột 1: đường dẫn file
+                'score': mos      # Cột 2: điểm số tương ứng
+            })
+            df.to_csv(f'/home4/tuanlha/EXpressiveTTS/dataRawProcess/05_data_extract/{playlist_name}/{episode}_wvmos.csv', index=False)
+        except Exception as e:
+            logger.error(f"Error processing episode {episode}: {e}")
+            continue
 
 def process_filter_by_rating(args, episode):
     csv_path = f'./05_data_extract/{args.playlist_name}/'
@@ -62,10 +84,10 @@ def process_filter_by_rating(args, episode):
     # From df_1, remove audio which MOS < 3.5
     df_1 = df_1[df_1['mos_pred'] < 3.5]['deg']
     df_1 = df_1.tolist()
-    df_2 = df_2[df_2['score']<3]['audio_path']
+    df_2 = df_2[df_2['score']<2.75]['audio_path']
     df_2 = df_2.tolist()
 
-    folder_path = os.path.join('./04_vad', args.playlist_name, episode)
+    folder_path = os.path.join('./04_denoise', args.playlist_name, episode)
 
     for file in df_1:
         file = os.path.join(folder_path, file)
